requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
loguru>=0.7.0
python-dotenv>=1.0.0
lxml>=4.9.0

python-whois>=0.8.0
dnspython>=2.3.0
